feat: merge ping and uptime cogs (#722)
